requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.57.0
accelerate>=0.20.0
datasets>=2.14.0
qwen-vl-utils>=0.0.8

# RL training
trl>=0.7.0
peft>=0.6.0

# Quantization
bitsandbytes>=0.41.0

# Image processing for SVG rendering
cairosvg>=2.7.0
pillow>=10.0.0

# Flash attention (optional but recommended for performance)
# Install separately with: pip install flash-attn --no-build-isolation
# flash-attn>=2.0.0

# Experiment tracking
wandb>=0.16.0

# Utilities
safetensors>=0.4.0
psutil>=5.9.0
requests>=2.31.0
